third_party/WebKit/LayoutTests/webaudio/offlineaudiocontext-suspend-resume-basic-expected.txt

Basic test for OfflineAudioContext.suspend() and OfflineAudioContext.resume().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS context.suspend() rejected correctly (with TypeError: Failed to execute 'suspend' on 'OfflineAudioContext': 1 argument required, but only 0 present.).
PASS context.suspend(-1.0) rejected correctly (with InvalidStateError: negative suspend time (-1) is not allowed).
PASS Scheduling a suspend in the past rejected correctly (with InvalidStateError: cannot schedule a suspend at frame 17536 (0.399229 seconds) because it is earlier than the current frame of 22016 (0.499229 seconds)).
PASS Calling multiple suspends at the same rendering quantum rejected correctly (with InvalidStateError: cannot schedule more than one suspend at frame 128 (0.00435374 seconds)).
PASS Resuming before suspend rejected correctly (with InvalidStateError: cannot resume a context that is not suspended).
PASS Calling resume on non-suspended context rejected correctly (with InvalidStateError: cannot resume a context that has not started).
PASS successfullyParsed is true

TEST COMPLETE
